Remove Cargo features from `wasmtime-internal-cranelift`
This commit removes the suite of Cargo features corresponding to wasm
proposals from the `wasmtime-internal-cranelift` crate. These features
are now instead unconditionally enabled for the crate. This is similarly
handled now in `wasmtime-internal-winch`.

The motivation for this commit is that all `#[cfg]` comes at a cost in
terms of CI, reading, writing, etc. The primary benefits of `#[cfg]` are
reducing runtime dependencies and reducing the compiled footprint of an
application. With `wasmtime-internal-cranelift` neither of these
benefits are realized with the current features because runtime
dependencies are the same with/without the features and Cranelift
dominates the compiled code size. Given the very-minor benefits, if any,
that are being earned it doesn't seem worth the cost.

This means that Cranelift translation unconditionally has support for
all wasm proposals, regardless of how Wasmtime is configured. Note
though that Wasmtime retains all these features just as before because
they do indeed gate actual runtime dependencies or significant chunks of
binary size.

prtest:full
