+++ crates/cranelift/src/bounds_checks.rs
@@ -53,7 +53,6 @@ pub enum BoundsCheck {
     /// ```ignore
     /// index + object_size <= bound
     /// ```
-    #[cfg(feature = "gc")]
     StaticObjectField {
         offset: u32,
         access_size: u8,
@@ -64,7 +63,6 @@ pub enum BoundsCheck {
     ///
     /// It is *your* responsibility to ensure that the `offset + access_size <=
     /// object_size` precondition holds.
-    #[cfg(feature = "gc")]
     DynamicObjectField {
         offset: ir::Value,
         object_size: ir::Value,
@@ -91,7 +89,6 @@ pub fn bounds_check_and_compute_addr(
             access_size,
         } => bounds_check_field_access(builder, env, heap, index, offset, access_size, trap),
 
-        #[cfg(feature = "gc")]
         BoundsCheck::StaticObjectField {
             offset,
             access_size,
@@ -129,7 +126,6 @@ pub fn bounds_check_and_compute_addr(
         // Compute the index of the end of the object, bounds check that and get
         // a pointer to just after the object, and then reverse offset from that
         // to get the pointer to the field being accessed.
-        #[cfg(feature = "gc")]
         BoundsCheck::DynamicObjectField {
             offset,
             object_size,
