+++ crates/cranelift/src/func_environ/gc/null.rs
